refactor(web): extract ClassicShell, make App a layout-agnostic container
Phase 2a of the sidebar redesign. App.svelte now owns all state, effects,
handlers and modals, and renders the page content as a snippet into a shell
component; ClassicShell owns the chrome (header, nav, filter bar, footer,
main wrapper). Svelte components emit no wrapper node, so the DOM under #app
is unchanged and the classic layout renders identically.

Adds lib/layout.ts (per-device localStorage layout setting, guarded against
blocked-storage SecurityError) as the foundation for the opt-in sidebar; the
shell switch and toggle UI land with the sidebar shell in a later phase.
