+++ packages/web/src/lib/layout.test.ts
@@ -1,0 +1,46 @@
+// @vitest-environment jsdom
+import { afterEach, describe, expect, it, vi } from 'vitest';
+import { getLayout, setLayout } from './layout';
+
+afterEach(() => {
+  localStorage.clear();
+  vi.restoreAllMocks();
+});
+
+describe('getLayout', () => {
+  it('defaults to classic when nothing is stored', () => {
+    expect(getLayout()).toBe('classic');
+  });
+
+  it('returns a stored valid layout', () => {
+    localStorage.setItem('inbox-rs:layout', 'sidebar');
+    expect(getLayout()).toBe('sidebar');
+  });
+
+  it('falls back to classic for an unrecognised value', () => {
+    localStorage.setItem('inbox-rs:layout', 'bogus');
+    expect(getLayout()).toBe('classic');
+  });
+
+  it('falls back to classic when storage access throws', () => {
+    vi.spyOn(Storage.prototype, 'getItem').mockImplementation(() => {
+      throw new DOMException('blocked', 'SecurityError');
+    });
+    expect(getLayout()).toBe('classic');
+  });
+});
+
+describe('setLayout', () => {
+  it('persists the chosen layout', () => {
+    setLayout('sidebar');
+    expect(localStorage.getItem('inbox-rs:layout')).toBe('sidebar');
+    expect(getLayout()).toBe('sidebar');
+  });
+
+  it('ignores blocked storage without throwing', () => {
+    vi.spyOn(Storage.prototype, 'setItem').mockImplementation(() => {
+      throw new DOMException('blocked', 'SecurityError');
+    });
+    expect(() => setLayout('sidebar')).not.toThrow();
+  });
+});
